requirements.txt
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.4.22
dotenv==0.9.9
exceptiongroup==1.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
ollama==0.6.1
pydantic==2.13.3
pydantic_core==2.46.3
python-dotenv==1.2.2
PyYAML==6.0.3
typing-inspection==0.4.2
typing_extensions==4.15.0
